move result files
